pyproject.toml
[project]
name = "flow-runner"
version = "0.2.0"
description = "YAML-configured flow engine with multi-provider LLM routing and eval"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
flow-run = "flow_runner.cli:main"
flow-prompt = "flow_runner.prompt_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
